Fix the order of Outputs, Derivatives and InitialUnknows in ModelStructure
FMI Compliance Checker complains

```
[ERROR][FMI2XML] [Line:36] XML element 'Derivatives' cannot be placed after element 'InitialUnknowns', skipping
[ERROR][FMI2XML] [Line:37] XML element 'Outputs' cannot be placed after element 'InitialUnknowns', skipping
```

which is caused by the re-sorting in the JSON loader for Python < 3.6. The issue vanishes if the object_pairs_hook is set to use the ordered dictionary type. See https://stackoverflow.com/a/47881325/8520615 for further information.

Additionally, The pretty-printing from lxml.etree is used instead the custom pretty-printing function.