Add support for scalar block parameters
Block parameters weren't supported yet. Unfortunately, the address of block parameters cannot be accessed by the name of the structure member (at least for R2015b) but need to get/set by the address map. Therefore, the index in the address map is part of the JSON API. Later, this index is removed, since it must not be part of the XML API.